Refactorización en migraciones

+++ database/migrations/2017_04_19_182816_add_fields_to_instructors_table.php
@@ -15,7 +15,8 @@ class AddFieldsToInstructorsTable extends Migration
     {
         Schema::table('instructors', function (Blueprint $table) {
             $table->string('address', 100)->after('phone_number');
-            $table->float('salary')->after('address');
+            $table->decimal('salary', 10, 2)->after('address');
+            //$table->float('salary')->after('address');
             //$table->integer('schedule_id')->unsigned()->index();
         });
     }
